requirements.txt
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
dateparser==1.1.1
distlib==0.3.4
Faker==13.3.2
filelock==3.6.0
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
linecache2==1.0.0
MarkupSafe==2.1.1
mock==4.0.3
Naked==0.1.31
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycryptodome==3.14.1
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-mock==3.7.0
pytest-xprocess==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
shellescape==3.8.1
six==1.16.0
tomli==2.0.1
traceback2==1.4.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.0.3
zipp==3.7.0
